Update and rename subroutines/define-and-call/square_root.cs to subroutines/define-and-call/c-sharp/square_root.cs
